requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
Cython==3.0.10
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
einops==0.8.0
einx==0.3.0
encodec==0.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.0
gne==0.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib_resources==6.4.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.77
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
openai==1.34.0
orjson==3.10.4
packaging==23.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
sympy==1.12.1
tenacity==8.3.0
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
vector-quantize-pytorch==1.14.24
vocos==0.1.0
wcwidth==0.2.13
WeTextProcessing==1.0.1
yarl==1.9.4
